requirements.txt
nemoguardrails>=0.9.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.108.0
uvicorn>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
